#16 fixed crac_protobuf name dependency and using a float for polling interval

+++ crac_server/component/telescope/telescope.py
@@ -155,7 +155,7 @@ class Telescope(ABC):
                 continue
             finally:
                 self.__disconnect()
-                sleep(config.Config.getInt("polling_interval", "telescope"))
+                sleep(config.Config.getFloat("polling_interval", "telescope"))
         else:
             self._reset()
             self.__disconnect()
